moves one function from app.py to a more dedicated file. Fix search request containing '
